fix error when creating secret using ssh auth
